"Fix" NPE with wither effect
For some reason, there is a mod that crashes with  minechem by pointing to PharmacologyEffectRegistry.java: and the line number containing "wither".
As far as I understand, some mod does something with the name "wither" and that's why it can't find it.
I just replaced the wither effect with its id